# contracts/operations/v1/openapi.yaml
openapi: 3.1.0
info:
  title: BoatOps Internal Operations API
  version: 1.0.0-alpha.5
  summary: Private operations commands and operator schedule catalog/calendar reads.
  description: >-
    Private alpha contract. Inventory and trip commands require operations.write.
    Finance writes require operations.finance.write, and finance reads require
    operations.finance.read or operations.finance.write. Schedule reads require
    operations.schedule.read or operations.schedule.write; schedule writes require
    operations.schedule.write. Schedule catalog writes use canonical server-owned
    identities and audit logs. This API is not part of the channel-facing Inventory
    Provider API and must not be exposed to untrusted channel clients.
servers:
  - url: https://boatops.ayany.com
    description: Ayany private alpha target
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/internal/v1/finance/accounts:
    post:
      tags: [Finance setup]
      operationId: createCashAccount
      summary: Create a payment-source account master
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateCashAccountRequest'}
      responses:
        '201':
          description: Cash or payment-source account created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SetupCommandResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /api/internal/v1/finance/expense-categories:
    post:
      tags: [Finance setup]
      operationId: createExpenseCategory
      summary: Create an operating-expense category
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateExpenseCategoryRequest'}
      responses:
        '201':
          description: Expense category created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SetupCommandResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/internal/v1/stock/items:
    post:
      tags: [Finance setup]
      operationId: createStockItem
      summary: Create a beverage, food, cleaning, safety, or maintenance stock item
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateStockItemRequest'}
      responses:
        '201':
          description: Stock item created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SetupCommandResponse'}
        '403': {$ref: '#/components/responses/AuthorizationFailed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque API token with explicit operations, finance, or schedule scope
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
        minLength: 36
        maxLength: 255
...
  responses:
    AuthorizationFailed:
      description: Authentication, organization isolation, or required operation scope failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    Conflict:
      description: Inventory, state, readiness, external-reference, or idempotency conflict.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
    ValidationFailed:
      description: Request validation failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
...
    CreateCashAccountRequest:
      type: object
      additionalProperties: false
      required: [external_reference, name, account_type, currency]
      properties:
        external_reference: {type: string, minLength: 1, maxLength: 255}
        name: {type: string, minLength: 1, maxLength: 255}
        account_type:
          type: string
          enum: [CASH, BANK, COMPANY_CARD, EMPLOYEE_ADVANCE, OTHER]
        currency: {type: string, pattern: '^[A-Z]{3}$'}
    CreateExpenseCategoryRequest:
      type: object
      additionalProperties: false
      required: [code, name, cost_scope]
      properties:
        code: {type: string, pattern: '^[A-Z0-9_-]+$', maxLength: 100}
        name: {type: string, minLength: 1, maxLength: 255}
        cost_scope:
          type: string
          enum: [DIRECT, BOAT_DAILY, COMMON]
    CreateStockItemRequest:
      type: object
      additionalProperties: false
      required: [external_reference, name, category, unit, currency]
      properties:
        external_reference: {type: string, minLength: 1, maxLength: 255}
        name: {type: string, minLength: 1, maxLength: 255}
        category:
          type: string
          enum: [BEVERAGE, FOOD, CLEANING, SAFETY, MAINTENANCE, OTHER]
        unit: {type: string, pattern: '^[A-Z0-9_-]+$', maxLength: 50}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        minimum_stock_quantity: {type: number, minimum: 0, multipleOf: 0.001}
...
    SetupCommandResponse:
      type: object
      additionalProperties: false
      required: [request_id, idempotency_key, organization_id, status, code, occurred_at, business_timezone]
      properties:
        request_id: {type: string, format: uuid}
        idempotency_key: {type: string}
        organization_id: {type: integer, minimum: 1}
        cash_account_id: {type: integer, minimum: 1}
        expense_category_id: {type: integer, minimum: 1}
        item_id: {type: integer, minimum: 1}
        external_reference: {type: string}
        category_code: {type: string}
        account_type: {type: string}
        cost_scope: {type: string}
        category: {type: string}
        unit: {type: string}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        minimum_stock_quantity: {type: string, pattern: '^\d+\.\d{3}$'}
        status: {const: ACTIVE}
        code:
          type: string
          enum: [CASH_ACCOUNT_CREATED, EXPENSE_CATEGORY_CREATED, STOCK_ITEM_CREATED]
        occurred_at: {type: string, format: date-time}
        business_timezone: {type: string}
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [request_id, code, message, retryable, manual_action_required]
      properties:
        request_id: {type: string, format: uuid}
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        manual_action_required: {type: boolean}